requirements.txt
asgiref==3.5.0
certifi==2021.10.8
configparser==5.2.0
Django
django-minio-backend==3.1.0b0
djangorestframework==3.13.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
djangorestframework_simplejwt
